feat: move `autoCast` getters to separate functions
Every type used to have a `.autoCast` and `.autoCastAll` getter that returns a clone of the type with autocast parsers added. This PR changes that to be separate functions with the same name. This helps with future changes, because maintaining getters that return the `this` types is becoming increasingly hard.

BREAKING CHANGE: The getters `.autoCast` and `.autoCastAll` are removed and moved to separately exported functions.